requirements.txt
pyTelegramBotAPI
Flask
Gunicorn
requests
quickchart.io